semua network query done kecuali lencanas

+++ Baksara/app/src/main/java/com/baksara/app/response/Langganan.kt
@@ -11,7 +11,7 @@ data class Langganan (
     @field: SerializedName("nama")
     val nama: String?,
     @field: SerializedName("harga")
-    val harga: Double?,
+    val harga: Float?,
     @field: SerializedName("durasi")
     val durasi: Int?,
 ) : Parcelable
